Build as Carthage dependency issues fixed
We are using RxAppState v.1.4.0 via Carthage.
It’s built successfully from time to time, but not stable. This issues need to be fixed for stable build.

And one more thing: I don’t know is it expected or not but it’s built twice when used via Carthage.
If it’s not expected, please check.